requirements.txt
django
djoser
djangorestframework_simplejwt
pillow
django-restframework
django-cors-headers
django-rest-swagger
drf-yasg
